Redo module mock and ensure --skip-formatting works as intended
